+++ src/lib/acespi/ace/acespi.hpp
@@ -12,14 +12,18 @@
 #include "pico/stdlib.h"
 
 /* GATAS. */
+#include "ace/coreutils.hpp"
 #include "ace/constants.hpp"
 #include "ace/basemodule.hpp"
 #include "ace/messages.hpp"
 #include "ace/semaphoreguard.hpp"
 
+#ifndef GATAS_SPI_DEFAULT_BUS_FREQUENCY
+#define GATAS_SPI_DEFAULT_BUS_FREQUENCY (15)
+#endif
+
 /**
- * Class that is responsible for managing the Single SPI bus between various devices
- * TODO: Add support for two SPI buses
+ * Class that is responsible for managing an SPI bus between various devices.
  */
 class AceSpi : public SpiModule, public etl::message_router<AceSpi>
 {
@@ -40,17 +44,26 @@ private:
     SemaphoreHandle_t mutex;
 public:
     static constexpr const etl::string_view NAME = "AceSpi";
-    AceSpi(etl::imessage_bus &bus, const GATAS::PinTypeMap &pins) : SpiModule(bus),
-                                                                      clk(pins.at(GATAS::PinType::CLK)),
-                                                                      mosi(pins.at(GATAS::PinType::MOSI)),
-                                                                      miso(pins.at(GATAS::PinType::MISO)),
-                                                                      rst(pins.at(GATAS::PinType::RST)),
-                                                                      spi(pins.at(GATAS::PinType::SPI)),
-                                                                      lastBusFrequency(GATAS_SPI_DEFAULT_BUS_FREQUENCY)
+    static constexpr uint8_t MAX_SPI_MODULES = SpiModule::MAX_SPI_MODULES;
+    static constexpr etl::array<etl::string_view, MAX_SPI_MODULES> NAMES{"AceSpi_0", "AceSpi_1"};
+
+    // Only fallback to AceSpi for AceSpi_0
+    static const GATAS::PinTypeMap pinMap(const Configuration &config, uint8_t device)
+    {
+        return config.pinMap(NAMES[device], device == 0 ? NAME : etl::string_view());
+    }
+
+    AceSpi(etl::imessage_bus &bus, const GATAS::PinTypeMap &pins, uint8_t device) : SpiModule(bus, SpiModule::NAMES[device]),
+                                                                                   clk(pins.at(GATAS::PinType::CLK)),
+                                                                                   mosi(pins.at(GATAS::PinType::MOSI)),
+                                                                                   miso(pins.at(GATAS::PinType::MISO)),
+                                                                                   rst(CoreUtils::getPin(pins, GATAS::PinType::RST, UINT8_MAX)),
+                                                                                   spi(pins.at(GATAS::PinType::SPI)),
+                                                                                   lastBusFrequency(GATAS_SPI_DEFAULT_BUS_FREQUENCY)
     {
     }
 
-    AceSpi(etl::imessage_bus &bus, const Configuration &config) : AceSpi(bus, config.pinMap(NAME))
+    AceSpi(etl::imessage_bus &bus, const Configuration &config, uint8_t device) : AceSpi(bus, pinMap(config, device), device)
     {
     }
 
@@ -76,7 +89,8 @@ public:
 
     virtual SpiGuard getLock(bool &locked) override;
 
-    virtual uint8_t spiNum() const {
+    virtual uint8_t spiNum() const override
+    {
         return spi;
     }