+++ src/lib/core/ace/coreutils.hpp
@@ -664,4 +664,20 @@ namespace CoreUtils
             return speedMs > GATAS::GROUNDSPEED_CONSIDERING_AIRBORN; // 10 m/s threshold for others
         }
     }
+
+    /**
+     * Find a pin number and if not found, return the default
+     * @param pinMap
+     * @param pin
+     * @param defaultNum
+     */
+    inline uint8_t getPin(const GATAS::PinTypeMap &pinMap, GATAS::PinType pin, uint8_t defaultNum)
+    {
+        auto it = pinMap.find(pin);
+        if (it != pinMap.end())
+        {
+            return static_cast<uint8_t>(it->second);
+        }
+        return defaultNum;
+    }
 }